+++ packages/unified-latex-to-pretext/libs/pre-conversion-subs/report-unsupported-macro-katex.ts
@@ -3,6 +3,7 @@ import { anyMacro, match } from "@unified-latex/unified-latex-util-match";
 import { visit } from "@unified-latex/unified-latex-util-visit";
 import { KATEX_SUPPORT } from "./katex-subs";
 import { VFileMessage } from "vfile-message";
+import { makeWarningMessage } from "./utils";
 
 /**
  * Return a list of macros used in ast that are unsupported by KaTeX
@@ -21,32 +22,16 @@ export function reportMacrosUnsupportedByKatex(ast: Ast.Ast): {
         if (anyMacro(node) && info.context.hasMathModeAncestor) {
             // check if not supported by katex
             if (!isSupported(node)) {
-                // get rid of type never
-                node = node as Ast.Macro;
-
-                // create a message specifing the unsupported macro
-                const message = new VFileMessage(
-                    `Warning: \"${node.content}\" is unsupported by Katex.`
+                // add a warning message
+                unsupported.messages.push(
+                    makeWarningMessage(
+                        node,
+                        `Warning: \"${
+                            (node as Ast.Macro).content
+                        }\" is unsupported by Katex.`,
+                        "report-unsupported-macro-katex"
+                    )
                 );
-
-                // add the position of the node if available
-                if (node.position) {
-                    message.line = node.position.start.line;
-                    message.column = node.position.start.column;
-                    message.position = {
-                        start: {
-                            line: node.position.start.line,
-                            column: node.position.start.column,
-                        },
-                        end: {
-                            line: node.position.end.line,
-                            column: node.position.end.column,
-                        },
-                    };
-                }
-
-                message.source = "latex-to-pretext:warning";
-                unsupported.messages.push(message);
             }
         }
     });
